+++ src/common/utils.h
@@ -33,6 +33,7 @@ std::vector<std::string> splitAny(const std::string& line,
 
 std::string join(const std::vector<std::string>& words, const std::string& del = " ");
 std::string join(const std::vector<size_t>& words, const std::string& del = " ");
+std::string join(const std::vector<float>& words, const std::string& del = " ", size_t prec = 5);
 
 std::string exec(const std::string& cmd, const std::vector<std::string>& args = {}, const std::string& arg = "");